update docs and GitHub Actions, reduce kamino setup-related duplication
