+++ .github/workflows/release-program.yaml
@@ -6,10 +6,9 @@ on:
       - "testest-program-*"
 
 env:
-  RUST_TOOLCHAIN: 1.75.0
-  SOLANA_CLI_VERSION: 1.18.17
-  ANCHOR_CLI_VERSION: 0.30.1
-  ANCHOR_SHA: e6d7dafe12da661a36ad1b4f3b5970e8986e5321
+  RUST_TOOLCHAIN: 1.90.0
+  SOLANA_CLI_VERSION: 3.1.11
+  ANCHOR_CLI_VERSION: 1.0.2
   CARGO_TERM_COLOR: always
 
 jobs:
