+++ .github/workflows/test.yaml
@@ -15,9 +15,8 @@ defaults:
 
 env:
   RUST_TOOLCHAIN: 1.90.0
-  SOLANA_CLI_VERSION: 2.2.1
-  ANCHOR_CLI_VERSION: 0.32.1
-  ANCHOR_SHA: e6d7dafe12da661a36ad1b4f3b5970e8986e5321
+  SOLANA_CLI_VERSION: 3.1.11
+  ANCHOR_CLI_VERSION: 1.0.2
   CARGO_TERM_COLOR: always
 
 concurrency:
@@ -58,9 +57,9 @@ jobs:
       with:
         path: |
           target/sbf
-        key: anchor-sbf-${{ runner.os }}-${{ env.ANCHOR_CLI_VERSION }}-${{ env.ANCHOR_SHA }}-${{ hashFiles('Anchor.toml', 'Cargo.lock', 'programs/**/Cargo.toml', 'programs/**/Xargo.toml', 'programs/**/src/**') }}
+        key: anchor-sbf-${{ runner.os }}-${{ env.ANCHOR_CLI_VERSION }}-${{ env.SOLANA_CLI_VERSION }}-${{ hashFiles('Anchor.toml', 'Cargo.lock', 'programs/**/Cargo.toml', 'programs/**/Xargo.toml', 'programs/**/src/**') }}
         restore-keys: |
-          anchor-sbf-${{ runner.os }}-${{ env.ANCHOR_CLI_VERSION }}-${{ env.ANCHOR_SHA }}-
+          anchor-sbf-${{ runner.os }}-${{ env.ANCHOR_CLI_VERSION }}-${{ env.SOLANA_CLI_VERSION }}-
     - name: Cache nextest host build
       uses: actions/cache@v3
       with:
@@ -165,7 +164,7 @@ jobs:
       run: echo "✅ Fuzzing completed"
 
   localnet-test-marginfi:
-    name: Anchor Bankrun Tests (Marginfi)
+    name: Anchor LiteSVM Tests (Marginfi)
     runs-on: ubuntu-latest
     needs: [ test-unit ]
     steps:
@@ -204,7 +203,7 @@ jobs:
       id: cache-transfer
       with:
         path: target/deploy/test_transfer_hook.so
-        key: anchor-transfer-hook-${{ runner.os }}-${{ env.ANCHOR_SHA }}-${{ hashFiles('programs/test_transfer_hook/**/*') }}
+        key: anchor-transfer-hook-${{ runner.os }}-${{ env.ANCHOR_CLI_VERSION }}-${{ env.SOLANA_CLI_VERSION }}-${{ hashFiles('programs/test_transfer_hook/**/*') }}
 
     - name: Build transfer hook
       if: steps.cache-transfer.outputs.cache-hit != 'true'
@@ -219,7 +218,7 @@ jobs:
           target/deploy/mocks.so
           target/idl/mocks.json
           target/types/mocks.ts
-        key: anchor-mocks-${{ runner.os }}-${{ env.ANCHOR_SHA }}-${{ hashFiles('programs/mocks/**/*') }}
+        key: anchor-mocks-${{ runner.os }}-${{ env.ANCHOR_CLI_VERSION }}-${{ env.SOLANA_CLI_VERSION }}-${{ hashFiles('programs/mocks/**/*') }}
 
     - name: Build mocks
       if: steps.cache-mocks.outputs.cache-hit != 'true'
