+++ guides/ADMIN/DEPLOY_GUIDE.md
@@ -4,7 +4,8 @@ Marginfi program authority is managed by squads (https://app.squads.so/squads/J3
 
 First you will need:
 
-- Solana tools 2.1.20 or later (`agave-install init 2.1.20`)
+- Agave/Solana tools 3.1.11 (`sh -c "$(curl -sSfL https://release.anza.xyz/v3.1.11/install)"`)
+- Anchor CLI 1.0.2 (`cargo install anchor-cli --version "1.0.2" --locked`)
 - solana-verify (`cargo install solana-verify`)
 - Docker (https://docs.docker.com/engine/install/ubuntu/)
 - A wallet with at least 15 SOL (this guide will assume your wallet is at `~/keys/mainnet-deploy.json`). Verify the pubkey of your wallet with `solana-keygen pubkey ~/keys/mainnet-deploy.json` and verify you have at least 15 SOL with `solana balance -k ~/keys/mainnet-deploy.json`
@@ -114,11 +115,12 @@ Update solana-verify (`cargo install solana-verify`) or manually edit cargo.lock
 version from 3 to 4
 
 ```
-Unable to find docker image for Solana version 2.1.20
-Using backup docker image for Solana version 2.1.13
+Unable to find docker image for Solana version 3.1.11
 ```
 
-No solution currently. The hash generated by 2.1.13 SHOULD BE the same as the hash generated by 2.1.20.
+Update `solana-verify` first. If no image exists for the exact toolchain version, do not silently
+substitute another Solana version for a production verified build; coordinate the expected hash with
+the other signers.
 
 Failing to verify build due to new multisig? Execute:
